pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0",
    "validators>=0.35.0",
]
